[IMP] l10n_it_edi, l10n_it_edi_ndd: merge module
This commit will merge the two modules.

task-4497665
