+++ addons/l10n_it_edi/data/l10n_it.document.type.csv
