+++ addons/l10n_it_edi/models/__init__.py
@@ -2,8 +2,10 @@
 from . import res_partner
 from . import res_company
 from . import res_config_settings
+from . import account_payment_method_line
 from . import account_move
 from . import account_move_send
 from . import ddt
 from . import ir_attachment
 from . import account_edi_proxy_user
+from. import l10n_it_document_type
