+++ addons/l10n_it_edi/models/l10n_it_document_type.py
