+++ addons/l10n_it_edi/security/ir.model.access.csv
@@ -1,2 +1,3 @@
 "id","name","model_id:id","group_id:id","perm_read","perm_write","perm_create","perm_unlink"
 "access_it_ddt_manager","it_ddt manager","model_l10n_it_ddt","account.group_account_invoice",1,1,1,1
+"access_l10n_it_document_type","access_l10n_it_document_type","model_l10n_it_document_type","account.group_account_invoice",1,1,1,1
