+++ addons/l10n_it_edi/tests/__init__.py
@@ -1,6 +1,7 @@
 # -*- coding: utf-8 -*-
 # Part of Odoo. See LICENSE file for full copyright and licensing details.
 from . import common
+from . import test_account_move_payment_method
 from . import test_edi_export
 from . import test_edi_import
 from . import test_edi_reverse_charge
