+++ addons/l10n_it_edi/views/account_payment_method.xml
