+++ addons/l10n_it_edi/views/l10n_it_document_type.xml
