--- addons/l10n_it_edi_ndd/__init__.py
@@ -1,1 +1,0 @@
-from . import models
