--- addons/l10n_it_edi_ndd/models/__init__.py
@@ -1,3 +1,0 @@
-from . import account_move
-from . import account_payment_methode_line
-from . import l10n_it_document_type
