--- addons/l10n_it_edi_ndd/security/ir.model.access.csv
@@ -1,2 +1,0 @@
-id,name,model_id:id,group_id:id,perm_read,perm_write,perm_create,perm_unlink
-l10n_it_edi_ndd.access_l10n_it_document_type,access_l10n_it_document_type,l10n_it_edi_ndd.model_l10n_it_document_type,account.group_account_invoice,1,1,1,1
