--- addons/l10n_it_edi_ndd/tests/__init__.py
@@ -1,1 +1,0 @@
-from . import test_account_move_payment_method
